requirements.txt
streamlit
google-genai
streamlit_option_menu
python-dotenv